Added support for rate-limit sessions cfg option
The frontend and listen templates were extended to support the
'rate-limit sessions <rate>' configuration option. If
`rate_limit_sessions` key is added to an entry in the
`haproxy_frontend` array with a value <rate>, the frontend definition
corresponding to the entry will limit the frontend to receiving at most
<rate> new sessions per second. Likewise, if the key is added
to an entry in the `haproxy_listen` array, the corresponding listener
will be rate limited in the same way.