adding support for SSL stats connections
The commit adds the ability to configure the HAProxy stats service to use TLS connections. When the new role variable `haproxy_stats_ssl` is set to `true` the `bind` configuration in the `listen stats` definition will append the TLS configuration parameters provided by the variable `haproxy_ssl`. To preserve the current behavior of the role, the default value of `haproxy_stats_ssl` is `false`.